Fix mouse problems when moving sequencer blocks (very common on OSX, but has also been observed on Linux). Plus a couple of minor clang compilation fixes.

+++ common/player.c
@@ -188,7 +188,7 @@ int PLAYER_get_block_delta_time(struct SeqTrack *seqtrack, STime time){
     }
     if(ret>=pc->reltime){
       RWarning("ret>pc->reltime: %d > %d",ret,pc->reltime);
-      return pc->reltime-1;
+      return (int)pc->reltime-1;
     }
     return ret;
   }else
